+++ src/basis/compiler/mcPrimitives.c
@@ -2,7 +2,7 @@
 #include "../../include/cfrtil.h"
 
 void
-MCP_MoveArrayN_To_Reg ( int32 reg, byte * array, int32 n )
+MCP_MoveArrayN_To_Reg ( int64 reg, byte * array, int64 n )
 {
     // pop reg, array, n
     // move array to reg
