+++ src/include/machineCodeMacros.h
@@ -89,6 +89,6 @@
 #define _Compile_TestImm_0_Rm( rm, sib, disp ) _Compile_Test ( MEM, 0, rm, disp, 0 )
 #endif
 #define  PUSH_IMM    0x68      
-#define _Compile_PushEspImm( data ) _Compile_InstructionX86 ( PUSH_IMM, 0, 0, 0, 0, 0, 0, ( int32 ) data, CELL )
+#define _Compile_PushEspImm( data ) _Compile_InstructionX86 ( PUSH_IMM, 0, 0, 0, 0, 0, 0, ( int64 ) data, CELL )
